If page is interactive, fire pageLoaded() immediately
When using Turbolinks with the `async` script attribute, Turbolinks may
execute after DOMContentLoaded. To fix this, we only attach a handler
to DOMContentLoaded if the readystate is not "loading" (that is to say,
it is "interactive" or "complete"). Otherwise, we fire the pageLoaded
event immediately without installing the handler.

This changes Turbolinks interaction with `defer` scripts. After this
change, an `async` Turbolinks script may execute *before* `defer`ed
scripts. This is because the browser executes all `defer` scripts *before*
the DOMContentLoaded event. Previously, Turbolinks would always execute
after these scripts, but may now execute before them in some cases
depending on when Turbolinks has finished downloading.