fix: unsupported browser (IE8)
`Turbolinks.visit` states:

If location is a cross-origin URL, or falls outside of the specified
root (see Setting a Root Location), or if the value of
Turbolinks.supported is false, Turbolinks performs a full page load by
setting window.location.

However this did not appear to be true for the unsupported browser IE8.

This commit allows `Turbolinks.visit` to gracefully fall back to the
standard javascript redirect via setting `window.location`